+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/main/scala/com/azure/cosmos/implementation/ChangeFeedSparkRowItem.scala
@@ -3,10 +3,12 @@
 
 package com.azure.cosmos.implementation
 
+import com.azure.cosmos.models.PartitionKey
 import org.apache.spark.sql.Row
 
 private[cosmos] case class ChangeFeedSparkRowItem
 (
   row: Row,
+  pkValue: Option[PartitionKey],
   lsn: String
 )